[FEAT, REFACTOR] 쿠폰 생성/조회/삭제 추가, PaymentDTO userID 자료형 수정

+++ src/main/java/com/kosa/backend/payment/dto/PaymentDTO.java
@@ -15,7 +15,7 @@ public class PaymentDTO {
     private String phoneNum;
     private String receiverName;
     private Long userId;
-    private Long couponId;
+    private Integer couponId;
     private String cardNumber;
     private String methodType;  // 추가: 카드인지 서드파티 페이인지 구분
     private String thirdPartyId; // 서드파티 페이용 ID
